refactor: move RangerAtlasConstants to auth-common and update imports
