+++ auth-agents-common/src/main/java/org/apache/atlas/plugin/util/ServiceDefUtil.java
@@ -19,7 +19,7 @@
 
 package org.apache.atlas.plugin.util;
 
-import org.apache.atlas.authorization.utils.RangerAtlasConstants;
+import org.apache.atlas.constants.RangerAtlasConstants;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.collections.MapUtils;
 import org.apache.commons.lang.StringUtils;
